requirements.txt
faster_whisper
python-ffmpeg
ffprobe-python
nvidia-cublas-cu12
nvidia-cudnn-cu12==9.*